pyproject.toml
[project]
name = "Aethon"
version = "1.0.0"
description = "Protecting precriptions, online reviews and articles free from fraud"
authors = [
    { name="Partha Khanna", email="parthakfree@gmail.com" },
    { name="Bhargav Shivakumar", email="bhargavshiv23@gmail.com" }
]
dependencies = [
    "requests>=2.31.0",
    "flask>=3.0.2",
    "flask-cors>=4.0.0",
    "pillow>=10.2.0",
    "pytesseract>=0.3.10",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.10"
